refactor: TimerMainView에서 StatusIndicator를 별도 컴포넌트로 분리
- StatusIndicator struct를 Components/StatusIndicator.swift로 추출
- TimerMainView 579줄 → 538줄로 축소
- 나머지 private 함수들은 screenVM 의존도가 높아 현재 구조 유지
